requirements.txt
Flask
psycopg2
python-dotenv
pandas 
numpy
flask-cors